Add server/IP ID caching

+++ rootfs/scripts/assign-floating-ip.sh
@@ -25,23 +25,33 @@ if [ -z "${HCLOUD_TOKEN}" ]; then
   exit 1
 fi
 
-SERVER_ID=$(
-  (curl -f -sSL \
-      --retry 2 --retry-delay 1 \
-      -H "Authorization: Bearer ${HCLOUD_TOKEN}" \
-      "https://api.hetzner.cloud/v1/servers?name=${HOSTNAME}" \
-    | jq .servers[0].id) \
-    2> >(tee -a /tmp/keepalived_notify.err >&2)
-)
+if [ -e /tmp/${HOSTNAME}_ID ]; then
+  SERVER_ID=$(cat /tmp/${HOSTNAME}_ID)
+else
+  SERVER_ID=$(
+    (curl -f -sSL \
+        --retry 2 --retry-delay 1 \
+        -H "Authorization: Bearer ${HCLOUD_TOKEN}" \
+        "https://api.hetzner.cloud/v1/servers?name=${HOSTNAME}" \
+      | jq .servers[0].id) \
+      2> >(tee -a /tmp/keepalived_notify.err >&2)
+  )
+  echo $SERVER_ID > /tmp/${HOSTNAME}_ID
+fi
 
-FLOATING_IP_ID=$(
-  (curl -f -sSL \
-      --retry 2 --retry-delay 1 \
-      -H "Authorization: Bearer ${HCLOUD_TOKEN}" \
-      "https://api.hetzner.cloud/v1/floating_ips" \
-    | jq ".floating_ips[] | select(.ip == \"${FLOATING_IP}\") | .id") \
-    2> >(tee -a /tmp/keepalived_notify.err >&2)
-)
+if [ -e /tmp/${FLOATING_IP}_ID ]; then
+  FLOATING_IP_ID=$(cat /tmp/${FLOATING_IP}_ID)
+else
+  FLOATING_IP_ID=$(
+    (curl -f -sSL \
+        --retry 2 --retry-delay 1 \
+        -H "Authorization: Bearer ${HCLOUD_TOKEN}" \
+        "https://api.hetzner.cloud/v1/floating_ips" \
+      | jq ".floating_ips[] | select(.ip == \"${FLOATING_IP}\") | .id") \
+      2> >(tee -a /tmp/keepalived_notify.err >&2)
+  )
+  echo $FLOATING_IP_ID > /tmp/${FLOATING_IP}_ID
+fi
 
 (curl -f -sSL -X POST \
     --retry 2 --retry-delay 1 \
